Nagios core code
this code file is the core code for Nagios installation. I retrieved this code from the Nagios source doe repository on GitHub. Since I am using the NEMS image for my Raspberry Pi I will most likely not need this code but I wanted to have it in the repository as a reference for the installation and to have better understanding of how the program works.