Updated Road Map
I added changes to the Determine Requirements section, adding all the necessary hardware and software tools that will be needed for this project. Also added changes to the Acquire Necessary Tools section, adding comments on components and software that have been acquired and those that have yet to be acquired.