Create Cacti Linux XML file
Created a file containing the XML code for the cacti install. Another file that I will need not actually need for the device as the program will be installed directly onto the Raspberry Pi but the code is a good reference point and I can edit as necessary if i discover any changes that I wish to make.