Cacti Install Code
Created this code file as a reference for the install code. I plan on installing directly to the raspberry pi but wanted to have this file in the repository to have a better understanding of the install code for this programs and to be able to make any changes to the code if necesary.